first unit tests passing after conversion of to_instance with decorator
